fix(eslint): Apply custom rules only to JS files
